+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.48](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.47...v1.1.48) (2020-08-13)
+
+**Note:** Version bump only for package chippr-bots
+
+
+
+
+
+## [1.1.47](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.46...v1.1.47) (2020-08-12)
+
+**Note:** Version bump only for package chippr-bots
+
+
+
+
+
+## [1.1.46](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.45...v1.1.46) (2020-08-12)
+
+**Note:** Version bump only for package chippr-bots
+
+
+
+
+
 ## [1.1.45](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.44...v1.1.45) (2020-05-05)
 
 **Note:** Version bump only for package chippr-bots